+++ static/js/cmdr.js
@@ -251,7 +251,7 @@ function renderTables() {
   }
 
   if (filterDW3) {
-    races = races.filter(r => !(r.tags || '').split(',').map(t => t.trim()).includes('DW3'));
+    races = races.filter(r => !(r.tags || '').split(',').map(t => t.trim()).includes('Remote'));
   }
 
   const types = [...new Set(stats.races.map(r => r.type))].sort();